+++ billiards.Unity/CueCamera.cs
@@ -36,7 +36,7 @@ public class CueCamera : MonoBehaviour
     // Height the cue view should reach when the player lifts the camera.
     public float cueRaisedHeight = 0.82f;
     // Minimum height maintained when the player drops the camera toward the cue.
-    public float cueLoweredHeight = 0.24f;
+    public float cueLoweredHeight = 0.3f;
     // Keep a small safety buffer from the butt of the cue so the camera never
     // retreats past the stick and always looks down the shaft.
     public float cueButtClearance = 0.12f;
@@ -354,17 +354,14 @@ public class CueCamera : MonoBehaviour
         float clothAnchorHeight = minimumCueHeight;
         float heightScale = Mathf.Lerp(1f, Mathf.Clamp(cueHeightClothScale, 0.1f, 1f), blend);
         height = clothAnchorHeight + (height - clothAnchorHeight) * heightScale;
-        float maxRailClamp = railHeight + Mathf.Max(0f, railClearance);
-        float cueRailClamp = Mathf.Lerp(maxRailClamp, railHeight, blend);
-        cueRailClamp = Mathf.Max(cueRailClamp, railHeight);
-        height = Mathf.Max(height, cueRailClamp);
+        float minRailHeight = railHeight + Mathf.Max(0f, railClearance);
+        height = Mathf.Max(height, minRailHeight);
 
         Vector3 focus = CueBall.position;
-        float minimumHeightBuffer = Mathf.Lerp(minimumHeightAboveFocus, 0f, blend);
-        float minimumHeightOffset = Mathf.Max(minimumHeightBuffer, height - focus.y);
+        float minimumHeightOffset = Mathf.Max(minimumHeightAboveFocus, height - focus.y);
         Vector3 lookTarget = GetCueAimLookTarget(focus, cueAimForward);
 
-        ApplyCameraAt(focus, cueAimForward, distance, height, minimumHeightOffset, lookTarget, cueRailClamp);
+        ApplyCameraAt(focus, cueAimForward, distance, height, minimumHeightOffset, lookTarget);
 
         Vector3 flatForward = new Vector3(cueAimForward.x, 0f, cueAimForward.z);
         if (flatForward.sqrMagnitude > 0.0001f)
@@ -589,7 +586,7 @@ public class CueCamera : MonoBehaviour
         float minimumHeightOffset = Mathf.Max(minimumHeightAboveFocus, height - focus.y);
         Vector3 lookTarget = focus + Vector3.up * Mathf.Max(0f, broadcastHeightPadding);
 
-        ApplyShortRailCamera(focus, forward, distance, height, minimumHeightOffset, lookTarget, minRailHeight);
+        ApplyShortRailCamera(focus, forward, distance, height, minimumHeightOffset, lookTarget);
     }
 
     private float ComputeBroadcastDistance(Vector3 focus, float height, Vector3 forward, Camera cam)
@@ -683,8 +680,7 @@ public class CueCamera : MonoBehaviour
         Quaternion rotation = Quaternion.Euler(0f, yaw, 0f);
         Vector3 forward = rotation * Vector3.forward;
         Vector3 lookTarget = focusPosition + forward * 5f;
-        float minRailHeight = railHeight + Mathf.Max(0f, railClearance);
-        ApplyShortRailCamera(focusPosition, forward, distance, height, minimumHeightAboveFocus, lookTarget, minRailHeight);
+        ApplyShortRailCamera(focusPosition, forward, distance, height, minimumHeightAboveFocus, lookTarget);
     }
 
     private void ApplyShortRailCamera(
@@ -697,8 +693,7 @@ public class CueCamera : MonoBehaviour
         Quaternion rotation = Quaternion.Euler(0f, yaw, 0f);
         Vector3 forward = rotation * Vector3.forward;
         Vector3 lookTarget = focusPosition + forward * 5f;
-        float minRailHeight = railHeight + Mathf.Max(0f, railClearance);
-        ApplyShortRailCamera(focusPosition, forward, distance, height, minimumHeightOffset, lookTarget, minRailHeight);
+        ApplyShortRailCamera(focusPosition, forward, distance, height, minimumHeightOffset, lookTarget);
     }
 
     private void ApplyShortRailCamera(
@@ -707,11 +702,10 @@ public class CueCamera : MonoBehaviour
         float distance,
         float height,
         float minimumHeightOffset,
-        Vector3 lookTarget,
-        float minRailHeight
+        Vector3 lookTarget
     )
     {
-        ApplyCameraAt(focusPosition, forward, distance, height, minimumHeightOffset, lookTarget, minRailHeight);
+        ApplyCameraAt(focusPosition, forward, distance, height, minimumHeightOffset, lookTarget);
 
         Vector3 pos = transform.position;
         pos.x = 0f;
@@ -724,8 +718,7 @@ public class CueCamera : MonoBehaviour
         float distance,
         float height,
         float minimumHeightOffset,
-        Vector3 lookTarget,
-        float minRailHeight
+        Vector3 lookTarget
     )
     {
         if (forward.sqrMagnitude < 0.0001f)
@@ -737,8 +730,8 @@ public class CueCamera : MonoBehaviour
         Vector3 desiredPosition = focusPosition - forward * distance + Vector3.up * height;
 
         float minimumHeight = focusPosition.y + Mathf.Max(minimumHeightAboveFocus, minimumHeightOffset);
-        float railClamp = Mathf.Max(minRailHeight, railHeight);
-        minimumHeight = Mathf.Max(minimumHeight, railClamp);
+        float minRailHeight = railHeight + Mathf.Max(0f, railClearance);
+        minimumHeight = Mathf.Max(minimumHeight, minRailHeight);
 
         // Prevent the camera from getting stuck behind walls or decorations by
         // nudging it toward the table if something blocks the line of sight.
@@ -761,22 +754,7 @@ public class CueCamera : MonoBehaviour
         }
 
         desiredPosition.y = Mathf.Max(desiredPosition.y, minimumHeight);
-        desiredPosition.y = Mathf.Max(desiredPosition.y, railClamp);
-
-        if (distance > 0f)
-        {
-            Vector3 toFocus = focusPosition - desiredPosition;
-            float forwardDistance = Vector3.Dot(toFocus, forward);
-            float requiredDistance = Mathf.Max(0.01f, distance);
-            if (forwardDistance < requiredDistance)
-            {
-                float adjust = requiredDistance - forwardDistance;
-                desiredPosition -= forward * adjust;
-            }
-        }
-
-        desiredPosition.y = Mathf.Max(desiredPosition.y, minimumHeight);
-        desiredPosition.y = Mathf.Max(desiredPosition.y, railClamp);
+        desiredPosition.y = Mathf.Max(desiredPosition.y, minRailHeight);
         transform.position = desiredPosition;
         transform.LookAt(lookTarget);
     }
