--- lib/americanBilliards.d.ts
@@ -1,5 +1,0 @@
-export class AmericanBilliards {
-  state: any;
-  constructor();
-  shotTaken(shot: any): any;
-}
