--- lib/nineBall.d.ts
@@ -1,5 +1,0 @@
-export class NineBall {
-  state: any;
-  constructor();
-  shotTaken(shot: any): any;
-}
