--- lib/poolUk8Ball.d.ts
@@ -1,6 +1,0 @@
-export class UkPool {
-  constructor(rules?: any);
-  state: any;
-  shotTaken(shot: any): any;
-  startBreak(): void;
-}
