--- webapp/src/pages/Games/AmericanBilliardsLobby.jsx
@@ -1,11 +1,0 @@
-import BilliardsLobbyBase from '../../components/billiards/BilliardsLobbyBase.jsx';
-
-export default function AmericanBilliardsLobby() {
-  return (
-    <BilliardsLobbyBase
-      title="American Billiards Lobby"
-      gamePath="/games/american-billiards"
-      initialVariant="american"
-    />
-  );
-}
