--- webapp/src/pages/Games/NineBallLobby.jsx
@@ -1,11 +1,0 @@
-import BilliardsLobbyBase from '../../components/billiards/BilliardsLobbyBase.jsx';
-
-export default function NineBallLobby() {
-  return (
-    <BilliardsLobbyBase
-      title="9-Ball Lobby"
-      gamePath="/games/nine-ball"
-      initialVariant="9ball"
-    />
-  );
-}
