+++ webapp/src/pages/Games/Snooker.jsx
@@ -17,22 +17,13 @@ import {
 import { FLAG_EMOJIS } from '../../utils/flagEmojis.js';
 import { UnitySnookerRules } from '../../../../src/rules/UnitySnookerRules.ts';
 import { useAimCalibration } from '../../hooks/useAimCalibration.js';
+import { useIsMobile } from '../../hooks/useIsMobile.js';
 import { isGameMuted, getGameVolume } from '../../utils/sound.js';
 import { getBallMaterial as getBilliardBallMaterial } from '../../utils/ballMaterialFactory.js';
 import {
   createCueRackDisplay,
   CUE_RACK_PALETTE
 } from '../../utils/createCueRackDisplay.js';
-import useTelegramBackButton from '../../hooks/useTelegramBackButton.js';
-import {
-  adjustCornerNotchDepth,
-  adjustSideNotchDepth,
-  centroidFromRing,
-  multiPolygonToShapes,
-  scaleMultiPolygon,
-  scaleMultiPolygonBy,
-  signedRingArea
-} from '../../components/billiards/geometry.js';
 import {
   WOOD_FINISH_PRESETS,
   WOOD_GRAIN_OPTIONS,
@@ -41,10 +32,157 @@ import {
   DEFAULT_WOOD_TEXTURE_SIZE,
   applyWoodTextures
 } from '../../utils/woodMaterials.js';
-import { MobilePortraitCameraRig, rad } from './simpleOrbitCamera';
-import { isWebGLAvailable } from '../../utils/webgl.js';
 
-const LEGACY_SRGB_ENCODING = Reflect.get(THREE, 'sRGBEncoding');
+function signedRingArea(ring) {
+  let area = 0;
+  for (let i = 0; i < ring.length - 1; i++) {
+    const [x1, y1] = ring[i];
+    const [x2, y2] = ring[i + 1];
+    area += x1 * y2 - x2 * y1;
+  }
+  return area * 0.5;
+}
+
+function multiPolygonToShapes(mp) {
+  const shapes = [];
+  if (!Array.isArray(mp)) return shapes;
+  mp.forEach((poly) => {
+    if (!Array.isArray(poly) || !poly.length) return;
+    const outerRing = poly[0];
+    if (!Array.isArray(outerRing) || outerRing.length < 4) return;
+    const outerPts = outerRing.slice(0, -1);
+    if (!outerPts.length) return;
+    const outerClockwise = signedRingArea(outerRing) < 0;
+    const orderedOuter = outerClockwise
+      ? outerPts.slice().reverse()
+      : outerPts;
+    const shape = new THREE.Shape();
+    shape.moveTo(orderedOuter[0][0], orderedOuter[0][1]);
+    for (let i = 1; i < orderedOuter.length; i++) {
+      shape.lineTo(orderedOuter[i][0], orderedOuter[i][1]);
+    }
+    shape.closePath();
+
+    for (let ringIndex = 1; ringIndex < poly.length; ringIndex++) {
+      const holeRing = poly[ringIndex];
+      if (!Array.isArray(holeRing) || holeRing.length < 4) continue;
+      const holePts = holeRing.slice(0, -1);
+      if (!holePts.length) continue;
+      const holeClockwise = signedRingArea(holeRing) < 0;
+      const orderedHole = holeClockwise
+        ? holePts
+        : holePts.slice().reverse();
+      const hole = new THREE.Path();
+      hole.moveTo(orderedHole[0][0], orderedHole[0][1]);
+      for (let i = 1; i < orderedHole.length; i++) {
+        hole.lineTo(orderedHole[i][0], orderedHole[i][1]);
+      }
+      hole.closePath();
+      shape.holes.push(hole);
+    }
+
+    shapes.push(shape);
+  });
+  return shapes;
+}
+
+function centroidFromRing(ring) {
+  if (!Array.isArray(ring) || !ring.length) {
+    return { x: 0, y: 0 };
+  }
+  let sumX = 0;
+  let sumY = 0;
+  let count = 0;
+  for (let i = 0; i < ring.length; i++) {
+    const pt = ring[i];
+    if (!Array.isArray(pt) || pt.length < 2) continue;
+    sumX += pt[0];
+    sumY += pt[1];
+    count++;
+  }
+  if (!count) {
+    return { x: 0, y: 0 };
+  }
+  return { x: sumX / count, y: sumY / count };
+}
+
+function scaleMultiPolygon(mp, scale) {
+  if (!Array.isArray(mp) || typeof scale !== 'number') {
+    return [];
+  }
+  const clampedScale = Math.max(0, scale);
+  return mp
+    .map((poly) => {
+      if (!Array.isArray(poly) || !poly.length) return null;
+      const outerRing = poly[0];
+      const centroid = centroidFromRing(outerRing);
+      const scaledPoly = poly
+        .map((ring) => {
+          if (!Array.isArray(ring) || !ring.length) return null;
+          return ring
+            .map((pt) => {
+              if (!Array.isArray(pt) || pt.length < 2) return null;
+              const dx = pt[0] - centroid.x;
+              const dy = pt[1] - centroid.y;
+              return [centroid.x + dx * clampedScale, centroid.y + dy * clampedScale];
+            })
+            .filter(Boolean);
+        })
+        .filter((ring) => Array.isArray(ring) && ring.length > 0);
+      if (!scaledPoly.length) return null;
+      return scaledPoly;
+    })
+    .filter((poly) => Array.isArray(poly) && poly.length > 0);
+}
+
+function scaleMultiPolygonBy(mp, scale) {
+  if (!Array.isArray(mp)) {
+    return [];
+  }
+  if (!Number.isFinite(scale) || Math.abs(scale - 1) < 1e-6) {
+    return mp;
+  }
+  return scaleMultiPolygon(mp, scale);
+}
+
+function adjustCornerNotchDepth(mp, centerZ, sz) {
+  if (!Array.isArray(mp) || !Number.isFinite(centerZ) || !Number.isFinite(sz)) {
+    return Array.isArray(mp) ? mp : [];
+  }
+  return mp.map((poly) =>
+    Array.isArray(poly)
+      ? poly.map((ring) =>
+          Array.isArray(ring)
+            ? ring.map((pt) => {
+                if (!Array.isArray(pt) || pt.length < 2) return pt;
+                const [x, z] = pt;
+                const deltaZ = z - centerZ;
+                const towardCenter = -sz * deltaZ;
+                if (towardCenter <= 0) return [x, z];
+                return [x, centerZ + deltaZ * CHROME_CORNER_NOTCH_CENTER_SCALE];
+              })
+            : ring
+        )
+      : poly
+  );
+}
+
+function adjustSideNotchDepth(mp) {
+  if (!Array.isArray(mp)) return Array.isArray(mp) ? mp : [];
+  return mp.map((poly) =>
+    Array.isArray(poly)
+      ? poly.map((ring) =>
+          Array.isArray(ring)
+            ? ring.map((pt) => {
+                if (!Array.isArray(pt) || pt.length < 2) return pt;
+                const [x, z] = pt;
+                return [x, z * CHROME_SIDE_NOTCH_DEPTH_SCALE];
+              })
+            : ring
+        )
+      : poly
+  );
+}
 
 const POCKET_VISUAL_EXPANSION = 1.05;
 const CHROME_CORNER_POCKET_RADIUS_SCALE = 1;
@@ -1213,11 +1351,8 @@ const createClothTextures = (() => {
     colorMap.generateMipmaps = true;
     colorMap.minFilter = THREE.LinearMipmapLinearFilter;
     colorMap.magFilter = THREE.LinearFilter;
-    if ('colorSpace' in colorMap && THREE.SRGBColorSpace) {
-      colorMap.colorSpace = THREE.SRGBColorSpace;
-    } else if ('encoding' in colorMap && LEGACY_SRGB_ENCODING !== undefined) {
-      colorMap.encoding = LEGACY_SRGB_ENCODING;
-    }
+    if ('colorSpace' in colorMap) colorMap.colorSpace = THREE.SRGBColorSpace;
+    else colorMap.encoding = THREE.sRGBEncoding;
     colorMap.needsUpdate = true;
 
     const bumpCanvas = document.createElement('canvas');
@@ -1655,11 +1790,8 @@ const createCarpetTextures = (() => {
     texture.minFilter = THREE.LinearMipMapLinearFilter;
     texture.magFilter = THREE.LinearFilter;
     texture.generateMipmaps = true;
-    if ('colorSpace' in texture && THREE.SRGBColorSpace) {
-      texture.colorSpace = THREE.SRGBColorSpace;
-    } else if ('encoding' in texture && LEGACY_SRGB_ENCODING !== undefined) {
-      texture.encoding = LEGACY_SRGB_ENCODING;
-    }
+    if ('colorSpace' in texture) texture.colorSpace = THREE.SRGBColorSpace;
+    else texture.encoding = THREE.sRGBEncoding;
 
     // bump map: derive from red base with extra fiber noise
     const bumpCanvas = document.createElement('canvas');
@@ -2777,11 +2909,8 @@ function makeClothTexture(
   const repeatY = baseRepeat * (PLAY_H / TABLE.H);
   texture.repeat.set(repeatX, repeatY);
   texture.anisotropy = 48;
-  if ('colorSpace' in texture && THREE.SRGBColorSpace) {
-    texture.colorSpace = THREE.SRGBColorSpace;
-  } else if ('encoding' in texture && LEGACY_SRGB_ENCODING !== undefined) {
-    texture.encoding = LEGACY_SRGB_ENCODING;
-  }
+  if ('colorSpace' in texture) texture.colorSpace = THREE.SRGBColorSpace;
+  else texture.encoding = THREE.sRGBEncoding;
   texture.minFilter = THREE.LinearMipMapLinearFilter;
   texture.magFilter = THREE.LinearFilter;
   texture.generateMipmaps = true;
@@ -2872,11 +3001,8 @@ function makeWoodTexture({
   texture.wrapS = texture.wrapT = THREE.RepeatWrapping;
   texture.repeat.set(repeatX, repeatY);
   texture.anisotropy = 8;
-  if ('colorSpace' in texture && THREE.SRGBColorSpace) {
-    texture.colorSpace = THREE.SRGBColorSpace;
-  } else if ('encoding' in texture && LEGACY_SRGB_ENCODING !== undefined) {
-    texture.encoding = LEGACY_SRGB_ENCODING;
-  }
+  if ('colorSpace' in texture) texture.colorSpace = THREE.SRGBColorSpace;
+  else texture.encoding = THREE.sRGBEncoding;
   texture.needsUpdate = true;
   return texture;
 }
@@ -3808,12 +3934,7 @@ function Table3D(
       ]);
     }
     const union = polygonClipping.union(...unionParts);
-    const adjusted = adjustCornerNotchDepth(
-      union,
-      cz,
-      sz,
-      CHROME_CORNER_NOTCH_CENTER_SCALE
-    );
+    const adjusted = adjustCornerNotchDepth(union, cz, sz);
     if (CHROME_CORNER_NOTCH_EXPANSION_SCALE === 1) {
       return adjusted;
     }
@@ -3847,7 +3968,7 @@ function Table3D(
     );
 
     const union = polygonClipping.union(circle, throat);
-    return adjustSideNotchDepth(union, CHROME_SIDE_NOTCH_DEPTH_SCALE);
+    return adjustSideNotchDepth(union);
   };
 
   const chromePlates = new THREE.Group();
@@ -4594,7 +4715,6 @@ function applyTableFinishToTable(table, finish) {
 function SnookerGame() {
   const mountRef = useRef(null);
   const rafRef = useRef(null);
-  useTelegramBackButton();
   const rules = useMemo(() => new UnitySnookerRules(), []);
   const [tableFinishId, setTableFinishId] = useState(() => {
     if (typeof window !== 'undefined') {
@@ -5027,7 +5147,6 @@ function SnookerGame() {
   const [err, setErr] = useState(null);
   const fireRef = useRef(() => {}); // set from effect so slider can trigger fire()
   const cameraRef = useRef(null);
-  const mobileCameraRigRef = useRef(null);
   const sphRef = useRef(null);
   const initialOrbitRef = useRef(null);
   const aimFocusRef = useRef(null);
@@ -5580,13 +5699,6 @@ function SnookerGame() {
   }, [frameState, hud.turn, hud.over]);
 
   useEffect(() => {
-    if (!isWebGLAvailable()) {
-      const message = 'WebGL is not supported or disabled on this device.';
-      setErr((prev) => prev ?? message);
-      setLoading(false);
-      setLoadingProgress(100);
-      return;
-    }
     const host = mountRef.current;
     if (!host) return;
     let loadTimer = null;
@@ -5606,12 +5718,7 @@ function SnookerGame() {
         powerPreference: 'high-performance'
       });
       renderer.useLegacyLights = false;
-      if ('outputColorSpace' in renderer && THREE.SRGBColorSpace) {
-        renderer.outputColorSpace = THREE.SRGBColorSpace;
-      }
-      if ('outputEncoding' in renderer && LEGACY_SRGB_ENCODING !== undefined) {
-        renderer.outputEncoding = LEGACY_SRGB_ENCODING;
-      }
+      renderer.outputColorSpace = THREE.SRGBColorSpace;
       renderer.toneMapping = THREE.ACESFilmicToneMapping;
       renderer.toneMappingExposure = 1.2;
       const devicePixelRatio = window.devicePixelRatio || 1;
@@ -5713,11 +5820,8 @@ function SnookerGame() {
         texture.minFilter = THREE.LinearFilter;
         texture.magFilter = THREE.LinearFilter;
         texture.anisotropy = 4;
-        if ('colorSpace' in texture && THREE.SRGBColorSpace) {
-          texture.colorSpace = THREE.SRGBColorSpace;
-        } else if ('encoding' in texture && LEGACY_SRGB_ENCODING !== undefined) {
-          texture.encoding = LEGACY_SRGB_ENCODING;
-        }
+        if ('colorSpace' in texture) texture.colorSpace = THREE.SRGBColorSpace;
+        else texture.encoding = THREE.sRGBEncoding;
         let offset = 0;
         return {
           texture,
@@ -5755,11 +5859,8 @@ function SnookerGame() {
         texture.minFilter = THREE.LinearFilter;
         texture.magFilter = THREE.LinearFilter;
         texture.anisotropy = 8;
-        if ('colorSpace' in texture && THREE.SRGBColorSpace) {
-          texture.colorSpace = THREE.SRGBColorSpace;
-        } else if ('encoding' in texture && LEGACY_SRGB_ENCODING !== undefined) {
-          texture.encoding = LEGACY_SRGB_ENCODING;
-        }
+        if ('colorSpace' in texture) texture.colorSpace = THREE.SRGBColorSpace;
+        else texture.encoding = THREE.sRGBEncoding;
         let pulse = 0;
         const createAvatarStore = () => ({
           image: null,
@@ -6510,31 +6611,24 @@ function SnookerGame() {
       const camera = new THREE.PerspectiveCamera(
         CAMERA.fov,
         aspect,
-        CAMERA.near,
-        CAMERA.far
-      );
-      const mobileCameraRig = new MobilePortraitCameraRig(camera, {
-        theta: rad(35),
-        phi: Math.PI,
-        radius: fitRadius(camera, STANDING_VIEW.margin)
-      });
-      mobileCameraRig.setViewport(host.clientWidth, host.clientHeight);
-      mobileCameraRigRef.current = mobileCameraRig;
-      const standingPhi = THREE.MathUtils.clamp(
-        STANDING_VIEW.phi,
-        CAMERA.minPhi,
-        CAMERA.maxPhi
-      );
-      const standingRadius = clamp(
-        fitRadius(camera, STANDING_VIEW.margin),
-        CAMERA.minR,
-        CAMERA.maxR
-      );
-      const sph = new THREE.Spherical(
-        standingRadius,
-        standingPhi,
-        Math.PI
-      );
+          CAMERA.near,
+          CAMERA.far
+        );
+        const standingPhi = THREE.MathUtils.clamp(
+          STANDING_VIEW.phi,
+          CAMERA.minPhi,
+          CAMERA.maxPhi
+        );
+        const standingRadius = clamp(
+          fitRadius(camera, STANDING_VIEW.margin),
+          CAMERA.minR,
+          CAMERA.maxR
+        );
+        const sph = new THREE.Spherical(
+          standingRadius,
+          standingPhi,
+          Math.PI
+        );
         cameraBoundsRef.current = {
           cueShot: { phi: initialCuePhi, radius: initialCueRadius },
           standing: { phi: standingPhi, radius: standingRadius }
@@ -6760,37 +6854,6 @@ function SnookerGame() {
         };
 
         const updateCamera = () => {
-          const mobileRig = mobileCameraRigRef.current;
-          if (
-            mobileRig &&
-            !topViewRef.current &&
-            !(cueGalleryStateRef.current?.active ?? false) &&
-            !(activeShotView && activeShotView.mode === 'action') &&
-            !pocketCameraStateRef.current
-          ) {
-            const now = performance.now();
-            const target = getDefaultOrbitTarget()
-              .clone()
-              .multiplyScalar(worldScaleFactor);
-            mobileRig.setViewport(host.clientWidth, host.clientHeight);
-            mobileRig.setTarget(target);
-            mobileRig.setTableDimensions(
-              PLAY_W * worldScaleFactor,
-              PLAY_H * worldScaleFactor
-            );
-            mobileRig.update(now);
-            const state = mobileRig.getState?.();
-            const currentSph = sphRef.current;
-            if (state && currentSph) {
-              currentSph.radius = state.radius;
-              currentSph.theta = state.phi;
-              currentSph.phi = Math.max(1e-3, Math.PI / 2 - state.theta);
-            }
-            camera.up.set(0, 1, 0);
-            camera.lookAt(target);
-            activeRenderCameraRef.current = camera;
-            return camera;
-          }
           let renderCamera = camera;
           let lookTarget = null;
           let broadcastArgs = {
@@ -8162,12 +8225,6 @@ function SnookerGame() {
       sph.radius = clampOrbitRadius(sph.radius);
       worldScaleFactor = WORLD_SCALE;
       world.scale.setScalar(worldScaleFactor);
-      if (mobileCameraRigRef.current) {
-        mobileCameraRigRef.current.setTableDimensions(
-          PLAY_W * worldScaleFactor,
-          PLAY_H * worldScaleFactor
-        );
-      }
       if (broadcastCamerasRef.current) {
         const rig = broadcastCamerasRef.current;
         const focusWorld = rig.defaultFocus
@@ -10544,12 +10601,6 @@ function SnookerGame() {
           // Update canvas dimensions when the window size changes so the table
           // remains fully visible.
           renderer.setSize(host.clientWidth, host.clientHeight);
-          if (mobileCameraRigRef.current) {
-            mobileCameraRigRef.current.setViewport(
-              host.clientWidth,
-              host.clientHeight
-            );
-          }
           const margin = Math.max(
             STANDING_VIEW.margin,
             topViewRef.current
@@ -10575,7 +10626,6 @@ function SnookerGame() {
           pocketCamerasRef.current.clear();
           pocketDropRef.current.clear();
           activeRenderCameraRef.current = null;
-          mobileCameraRigRef.current = null;
           cueBodyRef.current = null;
           tipGroupRef.current = null;
           try {
@@ -10616,7 +10666,6 @@ function SnookerGame() {
           if (loadTimer) {
             clearTimeout(loadTimer);
           }
-          loadingClearedRef.current = false;
         };
       } catch (e) {
         console.error(e);
@@ -10640,6 +10689,7 @@ function SnookerGame() {
     const slider = new SnookerPowerSlider({
       mount,
       value: powerRef.current * 100,
+      cueSrc: '/assets/snooker/cue.webp',
       labels: true,
       onChange: (v) => setHud((s) => ({ ...s, power: v / 100 })),
       onCommit: () => {
@@ -11113,5 +11163,15 @@ function SnookerGame() {
 }
 
 export default function NewSnookerGame() {
+  const isMobileOrTablet = useIsMobile(1366);
+
+  if (!isMobileOrTablet) {
+    return (
+      <div className="flex items-center justify-center w-full h-full p-4 text-center">
+        <p>This game is available on mobile phones and tablets only.</p>
+      </div>
+    );
+  }
+
   return <SnookerGame />;
 }
