--- webapp/src/pages/Games/SnookerLobby.jsx
@@ -1,13 +1,0 @@
-import BilliardsLobbyBase from '../../components/billiards/BilliardsLobbyBase.jsx';
-
-export default function SnookerLobby() {
-  return (
-    <BilliardsLobbyBase
-      title="3D Snooker Lobby"
-      gamePath="/games/snooker"
-      variantParamKey={null}
-      tableSizeOptions={null}
-      initialVariant={null}
-    />
-  );
-}
