--- webapp/src/pages/Games/UkEightBallLobby.jsx
@@ -1,11 +1,0 @@
-import BilliardsLobbyBase from '../../components/billiards/BilliardsLobbyBase.jsx';
-
-export default function UkEightBallLobby() {
-  return (
-    <BilliardsLobbyBase
-      title="8 Pool UK Lobby"
-      gamePath="/games/uk-eight-ball"
-      initialVariant="uk"
-    />
-  );
-}
